requirements.txt
backcall==0.2.0
click==8.0.3
cycler==0.11.0
dataclasses==0.8
decorator==5.1.0
entrypoints==0.3
Flask==2.0.2
freetype-py==2.2.0
gif2numpy==1.3
importlib-metadata==4.8.1
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jupyter-client==7.0.6
jupyter-core==4.9.1
kaitaistruct==0.9
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mss==6.1.0
nest-asyncio==1.5.1
numpy==1.19.5
numpy2gif==1.0
opencv-contrib-python==4.4.0.46
opencv-python==4.5.4.58
pandas==1.1.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
PILasOPENCV==2.7
Pillow==8.4.0
pkg_resources==0.0.0
prompt-toolkit==3.0.21
ptyprocess==0.7.0
Pygments==2.10.0
pyparsing==3.0.4
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
tornado==6.1
traitlets==4.3.3
typing-extensions==3.10.0.2
wcwidth==0.2.5
Werkzeug==2.0.2
zipp==3.6.0
